add edit form add validation jquery to layout add validation messages in framework

+++ ShopingCenter/ShopManagement.Infrastructure.EFCore/Migrations/20220413083706_initDataBase.cs
@@ -11,7 +11,7 @@ namespace ShopManagement.Infrastructure.EFCore.Migrations
                 name: "ProductCategories",
                 columns: table => new
                 {
-                    Id = table.Column<int>(type: "int", nullable: false)
+                    Id = table.Column<int>(type: "bigint", nullable: false)
                         .Annotation("SqlServer:Identity", "1, 1"),
                     Name = table.Column<string>(type: "nvarchar(255)", maxLength: 255, nullable: false),
                     Description = table.Column<string>(type: "nvarchar(500)", maxLength: 500, nullable: true),
